fix: Use specified CA cert for grpc
This had been skipping out of the function early if a client key
wasn't present.

I don't believe that's correct. If I've specified a CA cert because
the MPI server I'm trying to talk to is signed by a non-standard
CA (e.g. N1 devenv) it should work.

Silently skipping the CA is really confusing and leads to

> Failed to create connection" error="rpc error: code = Unavailable desc = connection error: desc = \"transport: authentication handshake failed: tls: failed to verify certificate: x509: certificate signed by unknown authority\"